refactor: 암호화 기능은 성격을 고려해 별도의 패키지로 분리
