+++ src/main/java/com/gomo/app/common/security/jwt/application/port/VerifyJwtPortIn.java
@@ -1,4 +1,4 @@
-package com.gomo.app.common.jwt.port;
+package com.gomo.app.common.security.jwt.application.port;
 
 public interface VerifyJwtPortIn {
