+++ src/main/java/com/gomo/app/core/member/application/usecase/ResetPasswordUseCase.java
@@ -1,8 +1,8 @@
 package com.gomo.app.core.member.application.usecase;
 
 import com.gomo.app.common.arch.ApplicationService;
-import com.gomo.app.common.jwt.port.VerifyJwtPortIn;
-import com.gomo.app.core.member.application.port.EncodePasswordPortOut;
+import com.gomo.app.common.security.encoder.application.port.EncodePasswordPortIn;
+import com.gomo.app.common.security.jwt.application.port.VerifyJwtPortIn;
 import com.gomo.app.core.member.domain.model.Email;
 import com.gomo.app.core.member.domain.model.Member;
 import com.gomo.app.core.member.domain.model.Password;
@@ -17,13 +17,13 @@ import lombok.RequiredArgsConstructor;
 public class ResetPasswordUseCase {
 
 	private final VerifyJwtPortIn verifyJwtPortIn;
-	private final EncodePasswordPortOut encodePasswordPortOut;
+	private final EncodePasswordPortIn encodePasswordPortIn;
 	private final MemberService memberService;
 
 	public void reset(String email, String newPassword, String temporaryToken) {
 		verifyJwtPortIn.validateToken(temporaryToken);
 		Member member = memberService.findByEmail(Email.of(email));
-		String encoded = encodePasswordPortOut.encode(Password.ofRaw(newPassword).getPassword());
+		String encoded = encodePasswordPortIn.encode(Password.ofRaw(newPassword).getPassword());
 		member.updatePassword(Password.ofEncoded(encoded));
 	}
 }
