+++ src/main/java/com/gomo/app/core/member/application/usecase/VerifyEmailCodeUseCase.java
@@ -1,7 +1,7 @@
 package com.gomo.app.core.member.application.usecase;
 
 import com.gomo.app.common.arch.ApplicationService;
-import com.gomo.app.common.jwt.port.GenerateJwtPortIn;
+import com.gomo.app.common.security.jwt.application.port.GenerateJwtPortIn;
 import com.gomo.app.support.auth.application.port.VerifyAuthCodePortIn;
 import com.gomo.app.support.logging.AuditLog;
