+++ src/main/java/com/gomo/app/support/auth/application/usecase/AuthenticateUseCase.java
@@ -3,7 +3,7 @@ package com.gomo.app.support.auth.application.usecase;
 import java.util.UUID;
 
 import com.gomo.app.common.arch.ApplicationService;
-import com.gomo.app.common.jwt.port.VerifyJwtPortIn;
+import com.gomo.app.common.security.jwt.application.port.VerifyJwtPortIn;
 import com.gomo.app.core.member.application.port.LoginMemberPortIn;
 import com.gomo.app.support.auth.application.port.dto.AuthTokenDto;
 import com.gomo.app.support.auth.domain.model.AuthToken;
